refactor: Dropdown 위치 변경
